Exemplos de manipulação de arquivos

+++ exemplo-matriz-4.py
@@ -6,7 +6,7 @@ for i in range(3):
 	#laço para criar as colunas 
 	for j in range(5):
 		#Vai adicionando as notas em cada linha x coluna
-		linha.append(int(input('Digite a nota [Linha '+str(i) +', Coluna ' + str(j) +']:')))
+		linha.append(float(input('Digite a nota [Linha '+str(i) +', Coluna ' + str(j) +']:')))
 	#adiciona a linha na matriz	
 	turma.append(linha)
 	print(turma)
